% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70_showplot.R
\name{showplot}
\alias{showplot}
\title{Presents a specific graphic determined by its name.}
\usage{
showplot(data, vars, diagram = "blank", output = "html")
}
\arguments{
\item{data}{Data.frame. Default dataset to use for plot. If not already a
data.frame, it should be first coerced to by [as.data.frame()].}

\item{vars}{Character. A variable within the dataset.}

\item{diagram}{Character. Specific graphic to be presented.}

\item{output}{Character. The type of output can be a 'html' file, the
'plots pane' or the 'console'.}
}
\value{
A html file that includes a or the code to produce it.
}
\description{
In order to present the graphic, the user must define a dataset, a
variable whitin this dataset and a compatible type of graphic.
Future work will include graphics that involve more than one variable.
}
\examples{
\dontrun{showplot(iris, "Petal.Length", "point graph")}
}
