% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50_longplot.R
\name{longplot}
\alias{longplot}
\title{Presents a longplot in a html output.}
\usage{
longplot(data, vars, label = TRUE)
}
\arguments{
\item{data}{Data.frame. Default dataset to use for plot. If not already a
data.frame, it should be first coerced to by [as.data.frame()].}

\item{vars}{Character. A specific variable within the dataset. Future work will allow to include a vector of variables.}

\item{label}{Logical. If `TRUE` the output includes labels that show the names of the graphics that are being displayed.}
}
\value{
A html file that includes a range of graphics suitable for this particular combination of variables.
}
\description{
A longplot is a range of suitable graphics that represent the relationship
within the values of one, or a limited number, of variables in a dataset. Each
graphic relates the values of all the selected variables and eventually the
row number in which  they appear.
}
\details{
In order to present the range of graphics, the user must define a dataset and
select one variable whitin it. Future work will include the ability to relate more
than one variable.
}
\examples{
\dontrun{longplot(iris, "Petal.Length")}
}
