% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60_wideplot.R
\name{wideplot}
\alias{wideplot}
\title{Presents a wideplot in a html output.}
\usage{
wideplot(data, dataclass = c("datetime", "logical", "ordered", "factor",
  "character", "numeric"), datetime = c("line graph",
  "point-to-point graph", "point graph", "binned heatmap"),
  logical = c("line graph", "point-to-point graph", "point graph",
  "tile plot", "bar graph"), ordered = c("line graph",
  "point-to-point graph", "point graph", "tile plot", "bar graph"),
  factor = c("line graph", "point-to-point graph", "point graph",
  "tile plot", "bar graph", "freq. reordered bar graph",
  "alphab. reordered bar graph"), numeric = c("line graph",
  "stepped line graph", "point graph", "color point graph", "histogram",
  "color bar graph", "normal qq plot"), character = c("line graph",
  "point-to-point graph", "point graph", "tile plot", "bar graph",
  "freq. reordered bar graph", "alphab. reordered bar graph"), ncol = 7,
  label = "FALSE")
}
\arguments{
\item{data}{Data.frame. Default dataset to use for plot. If not already
a data.frame, it should be first coerced to by [as.data.frame()].}

\item{dataclass}{Character. Types of data to be considered.}

\item{datetime}{Character. Graphics for datetime variables.}

\item{logical}{Character. Graphics for logical variables.}

\item{ordered}{Character. Graphics for ordered factor variables.}

\item{factor}{Character. Graphics for factor variables.}

\item{numeric}{Character. Graphics for numeric variables.}

\item{character}{Character. Graphics for character variables.}

\item{ncol}{Numeric. Number of columns A number between 3 and 10.}

\item{label}{Logical. If `TRUE` the output includes labels that show the
names of the graphics that are being displayed.}
}
\value{
A html file that includes a matrix of graphics. The variables of a
the dataset are first grouped by the type of data, then, each variable is
graphically represented into a range of different grap hics in one row of
the matrix.
}
\description{
A wideplot is a matrix of graphics were the graphics within each row
corresponds to graphical representations of each one of the variables
considered within a given dataset.
}
\examples{
\dontrun{wideplot(iris)}
\dontrun{wideplot(as.data.frame(HairEyeColor))}
}
